refactoring

+++ 6kyu/Simple string indices/C++/include/solution.hpp
@@ -5,6 +5,6 @@
 
 #include <string>
 
-int solve(const std::string& str, int index);
+int solve(const std::string& str, size_t index);
 
 #endif
